Workflow task functiona optimization

+++ modules/workflow/src/main/java/org/shaolin/bmdp/workflow/coordinator/ICoordinatorService.java
@@ -33,7 +33,7 @@ public interface ICoordinatorService {
 	 * 
 	 * @return
 	 */
-	int getTaskSize();
+	long getTaskSize();
 	
 	/**
 	 *
